仮キャラクター配置、地面実装
