+++ .gitignore
@@ -85,4 +85,4 @@ crashlytics-build.properties
 *.vsix
 
 
-Sato/[Tt]emp/
+Assets/Sato/[Tt]emp/
