+++ Assets/Sato/DirtTile.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 201a369dced04b4409ccf075c1d4abfe
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
