+++ Assets/Sato/GrassTile.prefab
@@ -62,7 +62,7 @@ MeshRenderer:
   m_RenderingLayerMask: 1
   m_RendererPriority: 0
   m_Materials:
-  - {fileID: 2100000, guid: bc5365ee772cac7429cb9d073e5b195a, type: 2}
+  - {fileID: 2100000, guid: 0312270a05dd6f74d94c4668632e340a, type: 2}
   m_StaticBatchInfo:
     firstSubMesh: 0
     subMeshCount: 0
