+++ Assets/Sato/Materials.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c96d9f9936b52d946ab50d459e9ec6ac
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
