+++ Assets/Sato/Materials/Dirt_Path_Basecolor.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7cbdf94b5a2d912488c6d860367a2e87
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
