+++ Assets/Sato/Materials/GrassGreen.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: bc5365ee772cac7429cb9d073e5b195a
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
