+++ Assets/Sato/Materials/Grass_Dense_Tint_01_Base_Basecolor_A.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0312270a05dd6f74d94c4668632e340a
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
