+++ Assets/Sato/Materials/Grass_Flower_Tint_02_FlowerRed_B_Basecolor.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 85208fdac8c7c414ebce2fad4d7277c7
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
