+++ Assets/Sato/Materials/OceanBlue.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 67b5e3c4af770e64e8ee30fab2bbad71
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
