+++ Assets/Sato/Scripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 51809db448ac5b74881964580ec9f29b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
