+++ Assets/Sato/Scripts/MapGenerator.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ecb76aa2a10aebc479b2294da8842706
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
